requirements.txt
pytesseract
pdf2image
openai
python-dotenv
rich
tqdm